third_party/blink/web_tests/fast/frames/sandboxed-iframe-forms-dynamic-expected.txt

CONSOLE ERROR: Blocked form submission to 'javascript:window.top.disallowedFormSubmitted();' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
  

PASS
